refactor(edit_project): Refactored edit project page code
