+++ messages/de.json
@@ -200,6 +200,7 @@
         "Delete All Licenses": "Alle Lizenzen löschen?",
         "Delete Attachment": "Löschen der Anlage",
         "Delete Component": "Löschen der Komponente",
+        "Delete Project": "",
         "Delete Release": "Löschen",
         "Delete Vulnerability": "Löschen der Schwachstelle",
         "Delete all License Information": "Alle Lizenzinformationen löschen",
@@ -418,6 +419,7 @@
         "Link Releases": "Link Releases",
         "Link To Project": "Link zum Projekt",
         "Link to Projects": "Link zu Projekten",
+        "Linked Packages": "",
         "Linked Releases": "Linked Releases",
         "Linked Releases and Projects": "Linked Releases und Projekte",
         "Logout": "Anmelden",
@@ -691,6 +693,7 @@
         "The release has been successfully linked to project": "Die Veröffentlichung <strong>(ReleaseName)</strong> wurde erfolgreich mit dem Projekt verbunden <strong>(Projektname)</strong>",
         "The verification of vulnerabilities will be changed to": "Überprüfung <strong>{number}</strong> Sicherheitslücken werden geändert",
         "There are some errors while creating project": "Es gibt einige Fehler bei der Erstellung von Projekten",
+        "There are some errors while updating project": "",
         "This currently only supports SPDX RDF/XML files with a uniq described top level node": "Dies unterstützt derzeit nur SPDX RDF/XML-Dateien mit einem uniq beschriebenen Top Level Knoten",
         "This function is meant to be followed by a new license import": "Diese Funktion soll durch einen neuen Lizenzimport verfolgt werden.",
         "This release contains": "Diese Veröffentlichung <strong>{name}</strong> enthält: ",
@@ -707,8 +710,8 @@
         "URL": "URL",
         "Unknown": "Unbekannt",
         "Update": "Aktualisierung",
-        "Update Project": "Projekt aktualisieren",
         "Update Component": "Update Component",
+        "Update Project": "Projekt aktualisieren",
         "Update Release": "Update Release",
         "Update Setting": "",
         "Update Vulnerability": "Schwachstelle aktualisieren",
@@ -770,6 +773,7 @@
         "You will be notified on changes of an item if you have the selected role in the changed item": "",
         "Your project is created": "Ihr Projekt wird erstellt",
         "Your request completed successfully": "",
+        "Your project is updated": "",
         "additional data": "zusätzliche Daten",
         "as of": "als",
         "authentication": "Authentisierung",
