+++ messages/en.json
@@ -428,6 +428,7 @@
         "Link Releases": "Link Releases",
         "Link To Project": "Link To Project",
         "Link to Projects": "Link to Projects",
+        "Linked Packages": "Linked Packages",
         "Linked Releases": "Linked Releases",
         "Linked Releases and Projects": "Linked Releases and Projects",
         "List View": "List View",
