+++ messages/es.json
@@ -200,6 +200,7 @@
         "Delete All Licenses": "Eliminar todas las licencias?",
         "Delete Attachment": "Suprímase el adjunto",
         "Delete Component": "Suprimir el componente",
+        "Delete Project": "",
         "Delete Release": "Eliminar la liberación",
         "Delete Vulnerability": "Eliminar la vulnerabilidad",
         "Delete all License Information": "Eliminar toda la información de licencia",
@@ -418,6 +419,7 @@
         "Link Releases": "Link Releases",
         "Link To Project": "Enlace al proyecto",
         "Link to Projects": "Enlace a Proyectos",
+        "Linked Packages": "",
         "Linked Releases": "Comunicados vinculados",
         "Linked Releases and Projects": "Comunicados y proyectos vinculados",
         "Logout": "Sign Out",
@@ -691,6 +693,7 @@
         "The release has been successfully linked to project": "La liberación <strong>{releaseName}</strong> has been successfully linked to project <strong>{proyectoName}</strong>",
         "The verification of vulnerabilities will be changed to": "La verificación de <strong>{número}</strong> vulnerabilidades se cambiarán",
         "There are some errors while creating project": "Hay algunos errores al crear proyecto",
+        "There are some errors while updating project": "",
         "This currently only supports SPDX RDF/XML files with a uniq described top level node": "Actualmente solo admite archivos SPDX RDF/XML con un nodo de nivel superior descrito uniq",
         "This function is meant to be followed by a new license import": "Esta función debe ser seguida por una nueva importación de licencias.",
         "This release contains": "Esta liberación <strong>{Nombre}</strong> contiene: ",
@@ -707,8 +710,8 @@
         "URL": "URL",
         "Unknown": "Desconocido",
         "Update": "Actualización",
-        "Update Project": "Proyecto de actualización",
         "Update Component": "Componente de actualización",
+        "Update Project": "Proyecto de actualización",
         "Update Release": "Actualización",
         "Update Setting": "",
         "Update Vulnerability": "Actualización de la vulnerabilidad",
@@ -770,6 +773,7 @@
         "You will be notified on changes of an item if you have the selected role in the changed item": "",
         "Your project is created": "Su proyecto es creado",
         "Your request completed successfully": "",
+        "Your project is updated": "",
         "additional data": "Datos adicionales",
         "as of": "a",
         "authentication": "autenticación",
