+++ messages/fr.json
@@ -200,6 +200,7 @@
         "Delete All Licenses": "Supprimer toutes les licences?",
         "Delete Attachment": "Supprimer la pièce jointe",
         "Delete Component": "Supprimer le composant",
+        "Delete Project": "",
         "Delete Release": "Supprimer la publication",
         "Delete Vulnerability": "Supprimer la vulnérabilité",
         "Delete all License Information": "Supprimer toutes les informations de licence",
@@ -418,6 +419,7 @@
         "Link Releases": "Sorties de liens",
         "Link To Project": "Lien vers le projet",
         "Link to Projects": "Lien vers des projets",
+        "Linked Packages": "",
         "Linked Releases": "Communiqués liés",
         "Linked Releases and Projects": "Communiqués et projets liés",
         "Logout": "Déconnexion",
@@ -691,6 +693,7 @@
         "The release has been successfully linked to project": "La libération <strong>{releaseName}</strong> a été lié avec succès au projet <strong>{nom du projet}</strong>",
         "The verification of vulnerabilities will be changed to": "La vérification <strong>{numéro}</strong> les vulnérabilités seront changées en",
         "There are some errors while creating project": "Il y a quelques erreurs lors de la création du projet",
+        "There are some errors while updating project": "",
         "This currently only supports SPDX RDF/XML files with a uniq described top level node": "Ceci ne prend actuellement en charge que les fichiers SPDX RDF/XML avec un nœud uniq décrit haut niveau",
         "This function is meant to be followed by a new license import": "Cette fonction doit être suivie d'une nouvelle importation de licence.",
         "This release contains": "Cette version <strong>{nom}</strong> contient: ",
@@ -707,8 +710,8 @@
         "URL": "URL",
         "Unknown": "Inconnu",
         "Update": "Mise à jour",
-        "Update Project": "Mettre à jour le projet",
         "Update Component": "Composante mise à jour",
+        "Update Project": "Mettre à jour le projet",
         "Update Release": "Mise à jour",
         "Update Setting": "",
         "Update Vulnerability": "Mettre à jour la vulnérabilité",
@@ -770,6 +773,7 @@
         "You will be notified on changes of an item if you have the selected role in the changed item": "",
         "Your project is created": "Votre projet est créé",
         "Your request completed successfully": "",
+        "Your project is updated": "",
         "additional data": "données supplémentaires",
         "as of": "à partir de",
         "authentication": "authentification",
