+++ messages/ja.json
@@ -200,6 +200,7 @@
         "Delete All Licenses": "すべてのライセンスを削除します。",
         "Delete Attachment": "添付ファイルの削除",
         "Delete Component": "コンポーネント削除",
+        "Delete Project": "",
         "Delete Release": "リリースの削除",
         "Delete Vulnerability": "Delete Vulnerability",
         "Delete all License Information": "すべてのライセンス情報を削除します。",
@@ -417,6 +418,7 @@
         "Link Releases": "リンクリリース",
         "Link To Project": "プロジェクトへのリンク",
         "Link to Projects": "Link to Projects",
+        "Linked Packages": "",
         "Linked Releases": "リンクされたリリース",
         "Linked Releases and Projects": "リンクされたリリースとプロジェクト",
         "Logout": "ログアウト",
@@ -690,6 +692,7 @@
         "The release has been successfully linked to project": "The release <strong>{releaseName}</strong> has been successfully linked to project <strong>{projectName}</strong>",
         "The verification of vulnerabilities will be changed to": "<strong>{number}</strong> 件の脆弱性の検証を変更",
         "There are some errors while creating project": "プロジェクトの作成中にいくつかのエラーが発生しました",
+        "There are some errors while updating project": "",
         "This currently only supports SPDX RDF/XML files with a uniq described top level node": "This currently only supports SPDX RDF/XML files with a uniq described top level node",
         "This function is meant to be followed by a new license import": "この関数は、新しいライセンスのインポートに従うことを目的としています。",
         "This release contains": "このリリース <strong>{name}</strong> には次の内容が含まれています:",
@@ -706,8 +709,8 @@
         "URL": "URL",
         "Unknown": "未知",
         "Update": "アップデート",
-        "Update Project": "プロジェクトを更新する",
         "Update Component": "コンポーネントの更新",
+        "Update Project": "プロジェクトを更新する",
         "Update Release": "Update Release",
         "Update Setting": "",
         "Update Vulnerability": "アップデートの脆弱性",
@@ -769,6 +772,7 @@
         "You will be notified on changes of an item if you have the selected role in the changed item": "",
         "Your project is created": "プロジェクトが作成されました",
         "Your request completed successfully": "",
+        "Your project is updated": "",
         "additional data": "追加データ",
         "as of": "as of",
         "authentication": "authentication",
