+++ messages/ko.json
@@ -200,6 +200,7 @@
         "Delete All Licenses": "모든 라이센스 삭제?",
         "Delete Attachment": "첨부파일",
         "Delete Component": "제거 구성",
+        "Delete Project": "",
         "Delete Release": "파일 삭제",
         "Delete Vulnerability": "취약점 삭제",
         "Delete all License Information": "모든 라이센스 정보 삭제",
@@ -417,6 +418,7 @@
         "Link Releases": "링크 출시",
         "Link To Project": "관련 링크",
         "Link to Projects": "프로젝트",
+        "Linked Packages": "",
         "Linked Releases": "링크드 릴리즈",
         "Linked Releases and Projects": "링크드 릴리즈 및 프로젝트",
         "Logout": "이름 *",
@@ -690,6 +692,7 @@
         "The release has been successfully linked to project": "공지사항 <strong>이름 *</strong> 프로젝트에 성공적으로 연결되었습니다. <strong>프로젝트 이름</strong>",
         "The verification of vulnerabilities will be changed to": "인증현황 <strong>이름 *</strong> 취약점이 변경됩니다.",
         "There are some errors while creating project": "프로젝트를 만드는 동안 몇 가지 오류가 있습니다.",
+        "There are some errors while updating project": "",
         "This currently only supports SPDX RDF/XML files with a uniq described top level node": "현재 SPDX RDF/XML 파일을 uniq 기술 최고 수준의 노드로 지원합니다.",
         "This function is meant to be followed by a new license import": "이 기능은 새로운 라이센스 수입에 의해 따를 수 없습니다.",
         "This release contains": "이 릴리스 <strong>이름 *</strong> 포함 : ",
@@ -706,8 +709,8 @@
         "URL": "사이트 맵",
         "Unknown": "이름 *",
         "Update": "(주)",
-        "Update Project": "프로젝트 업데이트",
         "Update Component": "업데이트 구성",
+        "Update Project": "프로젝트 업데이트",
         "Update Release": "업데이트",
         "Update Setting": "",
         "Update Vulnerability": "업데이트 취약성",
@@ -769,6 +772,7 @@
         "You will be notified on changes of an item if you have the selected role in the changed item": "",
         "Your project is created": "당신의 프로젝트는",
         "Your request completed successfully": "",
+        "Your project is updated": "",
         "additional data": "더 많은 데이터",
         "as of": "이름 *",
         "authentication": "인증 및 인증",
