+++ messages/pt-BR.json
@@ -200,6 +200,7 @@
         "Delete All Licenses": "Excluir todas as licenças?",
         "Delete Attachment": "Deletar Anexo",
         "Delete Component": "Excluir Componente",
+        "Delete Project": "",
         "Delete Release": "Excluir Liberação",
         "Delete Vulnerability": "Excluir Vulnerabilidade",
         "Delete all License Information": "Exclua todas as informações de licença",
@@ -417,6 +418,7 @@
         "Link Releases": "Link dos Releases",
         "Link To Project": "Link para o projeto",
         "Link to Projects": "Link para projetos",
+        "Linked Packages": "",
         "Linked Releases": "Releases vinculados",
         "Linked Releases and Projects": "Releases e Projetos Vinculados",
         "Logout": "Sair",
@@ -690,6 +692,7 @@
         "The release has been successfully linked to project": "A versão <strong>{releaseName}</strong> foi vinculada com sucesso ao projeto <strong>{projectName}</strong>",
         "The verification of vulnerabilities will be changed to": "A verificação de <strong>{number}</strong> vulnerabilidades será alterada para",
         "There are some errors while creating project": "Ocorreram alguns erros ao criar o projeto",
+        "There are some errors while updating project": "",
         "This currently only supports SPDX RDF/XML files with a uniq described top level node": "Atualmente, isso só suporta arquivos SPDX RDF/XML com um nó de nível superior uniq descrito",
         "This function is meant to be followed by a new license import": "Esta função deve ser seguida por uma nova importação de licença.",
         "This release contains": "Esta versão <strong>{name}</strong> contém:",
@@ -706,8 +709,8 @@
         "URL": "URL",
         "Unknown": "Desconhecido",
         "Update": "Atualização",
-        "Update Project": "Atualizar projeto",
         "Update Component": "Atualizar componente",
+        "Update Project": "Atualizar projeto",
         "Update Release": "Atualizar versão",
         "Update Setting": "",
         "Update Vulnerability": "Atualizar Vulnerabilidade",
@@ -769,6 +772,7 @@
         "You will be notified on changes of an item if you have the selected role in the changed item": "",
         "Your project is created": "Seu projeto foi criado",
         "Your request completed successfully": "",
+        "Your project is updated": "",
         "additional data": "Dados adicionais ",
         "as of": "A partir de",
         "authentication": "Authentication",
