+++ messages/vi.json
@@ -200,6 +200,7 @@
         "Delete All Licenses": "Xóa tất cả giấy phép?",
         "Delete Attachment": "Xóa tệp đính kèm",
         "Delete Component": "Xóa thành phần",
+        "Delete Project": "",
         "Delete Release": "Xóa bản phát hành",
         "Delete Vulnerability": "Xóa lỗ hổng",
         "Delete all License Information": "Xóa tất cả thông tin giấy phép",
@@ -417,6 +418,7 @@
         "Link Releases": "Liên kết phát hành",
         "Link To Project": "Liên kết đến dự án",
         "Link to Projects": "Liên kết tới các dự án",
+        "Linked Packages": "",
         "Linked Releases": "Các bản phát hành được liên kết",
         "Linked Releases and Projects": "Các bản phát hành và dự án được liên kết",
         "Logout": "Đăng xuất",
@@ -690,6 +692,7 @@
         "The release has been successfully linked to project": "Bản phát hành <strong>{releaseName}</strong> đã được liên kết thành công với dự án <strong>{projectName}</strong>",
         "The verification of vulnerabilities will be changed to": "Việc xác minh lỗ hổng <strong>{number}</strong> sẽ được thay đổi thành",
         "There are some errors while creating project": "Có một số lỗi khi tạo dự án",
+        "There are some errors while updating project": "",
         "This currently only supports SPDX RDF/XML files with a uniq described top level node": "Điều này hiện chỉ hỗ trợ các tệp SPDX RDF/XML với nút cấp cao nhất được mô tả uniq",
         "This function is meant to be followed by a new license import": "Chức năng này được thực hiện sau khi nhập giấy phép mới.",
         "This release contains": "Bản phát hành này <strong>{name}</strong> chứa:",
@@ -706,8 +709,8 @@
         "URL": "URL",
         "Unknown": "không xác định",
         "Update": "Cập nhật",
-        "Update Project": "Cập nhật dự án",
         "Update Component": "Cập nhật thành phần",
+        "Update Project": "Cập nhật dự án",
         "Update Release": "Cập nhật bản phát hành",
         "Update Setting": "",
         "Update Vulnerability": "Cập nhật lỗ hổng",
@@ -769,6 +772,7 @@
         "You will be notified on changes of an item if you have the selected role in the changed item": "",
         "Your project is created": "Dự án của bạn được tạo",
         "Your request completed successfully": "",
+        "Your project is updated": "",
         "additional data": "dữ liệu bổ sung",
         "as of": "kể từ",
         "authentication": "xác thực",
