+++ messages/zh-CN.json
@@ -200,6 +200,7 @@
         "Delete All Licenses": "删除所有许可证吗 ?",
         "Delete Attachment": "删除附件",
         "Delete Component": "删除组件",
+        "Delete Project": "",
         "Delete Release": "删除发布",
         "Delete Vulnerability": "删除漏洞",
         "Delete all License Information": "删除所有许可信息",
@@ -417,6 +418,7 @@
         "Link Releases": "链接发布",
         "Link To Project": "项目链接",
         "Link to Projects": "项目链接",
+        "Linked Packages": "",
         "Linked Releases": "链接版本",
         "Linked Releases and Projects": "链接的版本和项目",
         "Logout": "登出",
@@ -690,6 +692,7 @@
         "The release has been successfully linked to project": "版本<strong>{releaseName}</strong>已成功链接到项目<strong>{projectName}</strong>",
         "The verification of vulnerabilities will be changed to": "<strong>{number}</strong> 个漏洞的验证将更改为",
         "There are some errors while creating project": "创建项目时出现一些错误",
+        "There are some errors while updating project": "",
         "This currently only supports SPDX RDF/XML files with a uniq described top level node": "当前仅支持具有 uniq 描述的顶级节点的 SPDX RDF/XML 文件",
         "This function is meant to be followed by a new license import": "此功能将随后导入新的许可证 。",
         "This release contains": "此版本<strong>{name}</strong>包含：",
@@ -706,8 +709,8 @@
         "URL": "网址",
         "Unknown": "未知",
         "Update": "更新",
-        "Update Project": "更新項目",
         "Update Component": "更新组件",
+        "Update Project": "更新項目",
         "Update Release": "更新发布",
         "Update Setting": "",
         "Update Vulnerability": "更新漏洞",
@@ -769,6 +772,7 @@
         "You will be notified on changes of an item if you have the selected role in the changed item": "",
         "Your project is created": "您的项目已创建",
         "Your request completed successfully": "",
+        "Your project is updated": "",
         "additional data": "附加数据",
         "as of": "作为",
         "authentication": "验证",
